fix(customdatepicker): fix customDatePicker
Before the fix the user wasn't able to pick passed hours even if the date is in the future. For
example if today is July 1st 2pm you can't assign a task to July 1st 12pm, but my mistake was that
user also wasn't able to assign the task for anything earlier than the 2pm even if it's July 2nd.
After I noticed this mistake I swapped min Time and Day to a filtering system.